requirements.txt
python-dotenv
gunicorn
Flask
torch
transformers
pillow